# backend/docs/openapi.yml
openapi: 3.0.0
info:
  title: User API
  version: 1.0.0
  description: API documentation for the User service
servers:
  - url: http://localhost:3000
paths:
  /user:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                age:
                  type: number
                email:
                  type: string
      responses:
        201:
          description: User created successfully
        400:
          description: Bad request
  /user/{id}:
    get:
      summary: Get a user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the user to retrieve
      responses:
        200:
          description: User retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  age:
                    type: number
                  email:
                    type: string
        404:
          description: User not found
    put:
      summary: Update a user by ID
      parameters:
          - in: path
            name: id
            required: true
            schema:
              type: string
            description: The ID of the user to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                age:
                  type: number
                email:
                  type: string
      responses:
        200:
          description: User updated successfully
        404:
          description: User not found
    delete:
      summary: Delete a user by ID
      parameters:
          - in: path
            name: id
            required: true
            schema:
              type: string
            description: The ID of the user to delete
      responses:
          200:
            description: User deleted successfully
          404:
            description: User not found